bugfix: avoid request trace config dependency cycle
