+++ xllm/core/framework/config/CMakeLists.txt
@@ -48,6 +48,5 @@ cc_library(
     glog::glog
     nlohmann_json::nlohmann_json
     :platform
-    :request
     :util
 )
